chore(Componentize): rename internal module name to ComponentizeCreator
